cry/flagservice/src/app/requirements.txt
Flask
pycryptodome
uwsgi
